Address review comments:
* move hashing related things in 'fnv1a.h'
* remove unused base struct and inheritance tree
* rename the namespace to reflect usage of FNV-1a only
